refactor: remove Textarea component and integrate its functionality into DocumentTitleInput
