+++ client/src/components/TiptapBlockEditor.tsx
@@ -34,8 +34,9 @@ export default function TiptapBlockEditor({ pageId, className = "" }: Props) {
       <div className="mx-auto max-w-full border rounded-xl overflow-hidden">
         <div className="h-[calc(100vh-80px)] overflow-y-auto p-8 md:p-10 lg:p-12">
           <DocumentTitleInput
-            value={(rootBlock.content?.title as string) || ""}
+            value={rootBlock.content?.title || ""}
             onChange={handleUpdateTitle}
+            placeholder="Page Title"
           />
           <Separator className="my-4" />
           <TiptapWrapper>
